# 8288107: Auto-vectorization for integer min/max