--- a/test/hotspot/jtreg/compiler/lib/ir_framework/IRNode.java
+++ b/test/hotspot/jtreg/compiler/lib/ir_framework/IRNode.java
@@ -209,6 +209,10 @@ public class IRNode {
     public static final String VECTOR_UCAST_I2X = START + "VectorUCastI2X" + MID + END;
     public static final String VECTOR_REINTERPRET = START + "VectorReinterpret" + MID + END;
 
+    public static final String Max_I = START + "MaxI" + MID + END;
+    public static final String Min_V = START + "MinV" + MID + END;
+    public static final String Max_V = START + "MaxV" + MID + END;
+
     public static final String FAST_LOCK   = START + "FastLock" + MID + END;
     public static final String FAST_UNLOCK = START + "FastUnlock" + MID + END;