@Bhavana-Kilambi 2022-07-13
Hi, Thank you for your review. I have answered your questions below : 

**- Is MaxINode::Ideal transformation is required for auto-vectorization?**
No, it is not required but these transformations are good to have as MaxI node does not have fundamental transformations defined, for ex – optimizing patterns like Max(a, a). So I have added the same transformations as MinINode so that it handles the above pattern and a couple of others. I bundled this together with the auto-vectorization code as I felt it would be good to have some basic optimizations defined for MaxINode (MinINode already has a few), which could help improve overall performance of this patch for certain cases. I can put it in a separate PR if that's better.

**- I don't see how it will "generates these instructions only when the loop is vectorized and generates the usual cmp-cmove instructions when the loop is not vectorizable or when the max/min operations are called outside of the loop."**
I have performed testing mainly on aarch64 and x86 machines. With this patch, I could see vector equivalent instructions for min/max operations (as described in the commit message) for vectorizable loops. For other non-vectorizable or scalar versions, the MinI/MaxI nodes in the respective *.ad files translate to compare-conditional-move instructions.

- As for the code generation for the Arrays.copyOfRange() (from the comment on JDK-8039104), these are my findings on an x86_64 machine :
Without my patch, it generates a cmp-cmovl sequence for the “min” operation in copyOfRange() while with my patch, it generates a cmp-cmovg. The assembly listing is shown below for both the cases -

Without my patch :

  . . . . 
  0x00007fb3890b05ac:   mov    %ecx,%r11d
  0x00007fb3890b05af:   sub    %edx,%r11d                   ;*isub {reexecute=0 rethrow=0 return_oop=0}
                                                            ; - java.util.Arrays::copyOfRange@2 (line 3819)
  0x00007fb3890b05b2:   test   %r11d,%r11d
  0x00007fb3890b05b5:   jl     0x00007fb3890b0824           ;*ifge {reexecute=0 rethrow=0 return_oop=0}
                                                            ; - java.util.Arrays::copyOfRange@5 (line 3820)
  . . . . .
  0x00007fb3890b05d0:   mov    %rsi,%rcx
  0x00007fb3890b05d3:   mov    0xc(%rsi),%r8d               ; implicit exception: dispatches to 0x00007fb3890b083c
                                                            ;*arraylength {reexecute=0 rethrow=0 return_oop=0}
                                                            ; - java.util.Arrays::copyOfRange@51 (line 3823)
  0x00007fb3890b05d7:   mov    %edx,%r10d
  . . . . .
  0x00007fb3890b05e6:   mov    %r8d,%ebx
  0x00007fb3890b05e9:   sub    %edx,%ebx                    ;*isub {reexecute=0 rethrow=0 return_oop=0}
                                                            ; - java.util.Arrays::copyOfRange@53 (line 3823)
  **0x00007fb3890b05eb:   cmp    %r11d,%ebx**
  0x00007fb3890b05ee:   mov    %r11d,%ebp
  **0x00007fb3890b05f1:   cmovl  %ebx,%ebp                    ;*invokestatic min {reexecute=0 rethrow=0 return_oop=0}
                                                            ; - java.util.Arrays::copyOfRange@55 (line 3824)**


With my patch : 
  . . . . .
  0x00007efce90b10ac:   mov    %ecx,%r11d
  0x00007efce90b10af:   sub    %edx,%r11d                   ;*isub {reexecute=0 rethrow=0 return_oop=0}
                                                            ; - java.util.Arrays::copyOfRange@2 (line 3819)
  0x00007efce90b10b2:   test   %r11d,%r11d
  0x00007efce90b10b5:   jl     0x00007efce90b1324           ;*ifge {reexecute=0 rethrow=0 return_oop=0}
                                                            ; - java.util.Arrays::copyOfRange@5 (line 3820)
  . . . . . 
  0x00007efce90b10d0:   mov    %rsi,%rcx
  0x00007efce90b10d3:   mov    0xc(%rsi),%r8d               ; implicit exception: dispatches to 0x00007efce90b133c
                                                            ;*arraylength {reexecute=0 rethrow=0 return_oop=0}
                                                            ; - java.util.Arrays::copyOfRange@51 (line 3823)
  0x00007efce90b10d7:   mov    %edx,%r10d
  . . . . .
  0x00007efce90b10e6:   mov    %r8d,%ebp
  0x00007efce90b10e9:   sub    %edx,%ebp
  **0x00007efce90b10eb:   cmp    %r11d,%ebp
  0x00007efce90b10ee:   cmovg  %r11d,%ebp                   ;*invokestatic min {reexecute=0 rethrow=0 return_oop=0}
                                                            ; - java.util.Arrays::copyOfRange@55 (line 3824)**

Although I suspected there should not be any degradation with my patch, I tried to do a quick bench-marking of this testcase with and without my patch to confirm, and here are the results - 
With my patch -
ByteArrMinMax.copyRange      2048       0  avgt   30  8.297 ± 0.305  ns/op

Without my patch - 
ByteArrMinMax.copyRange      2048       0  avgt   30  8.446 ± 0.105  ns/op

There isn't much difference in performance between the two cases. 

The above tests were run with -XX:-TieredCompilation flag to ensure the copyOfRange is being compiled by c2.