@vnkozlov 2022-07-13
Thank you for checking Arrays.copyOfRange() code generation.
I will start our benchmarks testing without **MaxINode::Ideal**