@vnkozlov 2022-07-15
Performance results (specjvm2008, dacapo, renaissance) on linux-x64 are neutral, just small variations.
So it is good.