@Bhavana-Kilambi 2022-07-15
Added a new commit with the MaxINode::Ideal tests related code stripped off and only retaining the code related to generating MinI/MaxI node for Math/min/max intrinsics.