@TobiHartmann 2022-10-24
This introduced a regression: [JDK-8294816](https://bugs.openjdk.org/browse/JDK-8294816) @Bhavana-Kilambi, could you please have a look?