@Bhavana-Kilambi 2022-10-28
Hello, I have debugged the problem and the issue seems to be with the wrong datatype being considered by the MinV/MaxV nodes during autovectorization. The integer inputs to the min/max operation are being downconverted to "short" (thereby losing the information present in higher order bits) leading to incorrect result. Plus Java Math API for min/max operations does not support short/byte subword types. I am working on a patch to fix this and will raise a PR soon. Thank you.